SpinGlassExhaustive

Index

SpinGlassPEPS.SpinGlassExhaustive.energy_qubo

SpinGlassPEPS.SpinGlassExhaustive.exhaustive_search

SpinGlassPEPS.SpinGlassExhaustive.exhaustive_search_bucket

SpinGlassPEPS.SpinGlassExhaustive.generate_random_graph

SpinGlassPEPS.SpinGlassExhaustive.get_energy_offset

SpinGlassPEPS.SpinGlassExhaustive.graph_to_dict

SpinGlassPEPS.SpinGlassExhaustive.graph_to_qubo

SpinGlassPEPS.SpinGlassExhaustive.kernel

SpinGlassPEPS.SpinGlassExhaustive.kernel_bucket

SpinGlassPEPS.SpinGlassExhaustive.kernel_qubo

SpinGlassPEPS.SpinGlassExhaustive.max_chunk_size

SpinGlassPEPS.SpinGlassExhaustive.naive_energy_kernel

SpinGlassPEPS.SpinGlassExhaustive.energy_qubo — Method

energy_qubo(state_code, graph)

state_code: state code for which the energy expressed in qubo is to be calculated.

graph: graph of the ising model.

Returns the state energy expressed as QUBO.

SpinGlassPEPS.SpinGlassExhaustive.exhaustive_search — Method

exhaustive_search(ig)

ig::IsingGraph: graph of ising model represented by IsingGraph structure.

Returns energies and states for provided graph by brute-forece alorithm based on GPU.

SpinGlassPEPS.SpinGlassExhaustive.exhaustive_search_bucket — Function

exhaustive_search_bucket(ig)
exhaustive_search_bucket(ig, how_many)

ig::IsingGraph: graph of ising model represented by IsingGraph structure.

Returns energies and states for provided graph by brute-forece alorithm supported by bucket selection based on GPU.

SpinGlassPEPS.SpinGlassExhaustive.generate_random_graph — Method

generate_random_graph(d)

d::Int: size of random graph.

Returns random array of size d.

SpinGlassPEPS.SpinGlassExhaustive.get_energy_offset — Method

get_energy_offset(graph)

graph: graph of the ising model.

Returns offest between Ising model graph and its QUBO representation.

SpinGlassPEPS.SpinGlassExhaustive.graph_to_dict — Method

graph_to_dict(graph)

graph: graph of the ising model.

Converts Ising model graph to Dict.

SpinGlassPEPS.SpinGlassExhaustive.graph_to_qubo — Method

graph_to_qubo(graph)

graph: graph of the ising model.

Converts Ising model graph to QUBO representation.

SpinGlassPEPS.SpinGlassExhaustive.kernel — Method

kernel(graph, energies)

graph: graph of the ising model.

energies: array filled with zeros. Each array index represents the state of the system.

Returns energies for every state.

SpinGlassPEPS.SpinGlassExhaustive.kernel_bucket — Method

kernel_bucket(graph, energies, state_offset)

graph: graph of the ising model.

energies: array filled with zeros. Each array index represents the state of the system.

idx: list for collecting partial energy results.

Returns energies for given indexes.

SpinGlassPEPS.SpinGlassExhaustive.kernel_qubo — Method

kernel_qubo(graph, energies)

graph: graph of the ising model.

energies: array filled with zeros. Each array index represents the state of the system.

Returns energies energy expressed as QUBO for every state.

SpinGlassPEPS.SpinGlassExhaustive.max_chunk_size — Method

max_chunk_size()

Returns the maximum chunk size for the algorithm supported by bucket selection.

SpinGlassPEPS.SpinGlassExhaustive.naive_energy_kernel — Method

naive_energy_kernel(J, energies, σ)

J: diagonal elements of graph of the ising model..

energies: array filled with zeros. Each array index represents the state of the system.

σ: non-diagonal elements of of graph of the ising model.

Returns the state energy.

SpinGlassPEPS.SpinGlassNetworks.brute_force — Method

brute_force(ig, ; num_states)

ig::IsingGraph: graph of ising model represented by IsingGraph structure.

Returns energies and states for provided model by naive brute-forece alorithm based on GPU.

SpinGlassPEPS.SpinGlassNetworks.energy — Method

energy(state_code, graph)

state_code: state code for which the energy.

graph: graph of the ising model.

Returns the state energy.