Belief propagation

Local dimensional reduction can be achieved by selectively choosing states in the clustered Hamiltonian that have the lowest local energy in the cluster. This approach aims to reduce the dimensionality of the problem by focusing on the most relevant and energetically favorable states. It can be done by truncation based on energy or truncation based on Loopy Belief Propagation.

SpinGlassNetworks.clustered_hamiltonian_2site — Function

clustered_hamiltonian_2site(
    cl_h::LabelledGraphs.LabelledGraph{S, T},
    beta::Real
) -> LabelledGraphs.LabelledGraph{MetaGraphs.MetaDiGraph{Int64, Float64}}

Constructs a clustered Hamiltonian for a given clustered Hamiltonian with a 2-site cluster approximation used in Pegasus graph.

Arguments:

cl_h::LabelledGraph{S, T}: The clustered Hamiltonian represented as a labelled graph.

beta::Real: The inverse temperature parameter for the 2-site cluster Hamiltonian construction.

Returns:

new_cl_h::LabelledGraph{MetaDiGraph}: A new labelled graph representing the 2-site cluster Hamiltonian.

This function constructs a clustered Hamiltonian cl_h by applying a 2-site cluster approximation. It combines and merges vertices and edges of the original graph to create a simplified representation of the Hamiltonian.

The resulting new_cl_h graph represents the 2-site cluster Hamiltonian with simplified interactions between clusters. The energy values, projectors, and spectra associated with the new vertices and edges are computed based on the provided temperature parameter beta.

SpinGlassNetworks.belief_propagation — Function

belief_propagation(
    cl_h::LabelledGraphs.LabelledGraph{S, T},
    beta::Real;
    tol,
    iter
) -> Dict{Any, Any}

Perform loopy belief propagation on a given clustered Hamiltonian.

Arguments:

cl_h::LabelledGraph{S, T}: The clustered Hamiltonian represented as a labelled graph.

beta::Real: The inverse temperature parameter for the belief propagation algorithm.

tol::Real (optional, default=1e-6): The convergence tolerance. The algorithm stops when the message updates between iterations are smaller than this value.

iter::Int (optional, default=1): The maximum number of iterations to perform.

Returns:

beliefs::Dict: A dictionary where keys are vertices of clustered hamiltonian, and values are the

resulting beliefs after belief propagation.

The function implements loopy belief propagation on the given clustered hamiltonian cl_h to calculate beliefs for each vertex. Belief propagation is an iterative algorithm that computes beliefs by passing messages between vertices and edges of the clustered hamiltonian. The algorithm continues until convergence or until the specified maximum number of iterations is reached. The beliefs are computed based on the inverse temperature parameter beta, which controls the influence of energy values on the beliefs.

SpinGlassNetworks.truncate_clustered_hamiltonian_2site_BP — Function

truncate_clustered_hamiltonian_2site_BP(
    cl_h::LabelledGraphs.LabelledGraph{S, T},
    beliefs::Dict,
    num_states::Int64;
    ...
) -> LabelledGraphs.LabelledGraph{MetaGraphs.MetaDiGraph{Int64, Float64}}
truncate_clustered_hamiltonian_2site_BP(
    cl_h::LabelledGraphs.LabelledGraph{S, T},
    beliefs::Dict,
    num_states::Int64,
    result_folder::String;
    ...
) -> LabelledGraphs.LabelledGraph{MetaGraphs.MetaDiGraph{Int64, Float64}}
truncate_clustered_hamiltonian_2site_BP(
    cl_h::LabelledGraphs.LabelledGraph{S, T},
    beliefs::Dict,
    num_states::Int64,
    result_folder::String,
    inst::String;
    beta
) -> LabelledGraphs.LabelledGraph{MetaGraphs.MetaDiGraph{Int64, Float64}}

Truncate a clustered Hamiltonian based on 2-site belief propagation states.

This function truncates a clustered Hamiltonian by considering 2-site belief propagation states and selecting the most probable states to keep. It computes the beliefs for all 2-site combinations and selects the states that maximize the probability.

Arguments:

cl_h::LabelledGraph{S, T}: The clustered Hamiltonian represented as a labelled graph.

beliefs::Dict: A dictionary containing belief values for 2-site interactions.

num_states::Int: The maximum number of most probable states to keep.

beta::Real (optional): The inverse temperature parameter (default is 1.0).

Returns:

LabelledGraph{S, T}: A truncated clustered Hamiltonian.